Modify emma
