+++ themes/gatsby-theme-emma/src/@lekoarts/gatsby-theme-emma-core/components/project.tsx
@@ -1,0 +1,15 @@
+import React from "react"
+import Project from "../../../components/project"
+
+type Props = {
+  data: {
+    project: any
+    [key: string]: any
+  }
+}
+
+export default ({ data }: Props) => {
+  const { project } = data
+
+  return <Project data={{ ...data, project }} />
+}
