+++ themes/gatsby-theme-emma/src/@lekoarts/gatsby-theme-emma-core/components/projects.tsx
@@ -1,0 +1,15 @@
+import React from "react"
+import Projects from "../../../components/projects"
+
+type Props = {
+  data: {
+    allProject: any
+    [key: string]: string
+  }
+}
+
+export default ({ data }: Props) => {
+  const { allProject } = data
+
+  return <Projects projects={allProject.nodes} />
+}
